requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.1
category_encoders==2.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
db-dtypes==1.4.1
exceptiongroup==1.2.2
fastapi==0.115.8
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
